requirements.txt
django~=5.1.3
djangorestframework~=3.15.2
environs
drf-spectacular~=0.27.2